requirements.txt
asyncio
faker
fastapi
pylint
pytest
python-dotenv
requests
responses
uvicorn
wheel